requirements.txt
python_telegram_bot==20.0
requests==2.28.2